- prevent editing of locked applications
