+++ pyramus/src/main/java/fi/otavanopisto/pyramus/views/applications/EditApplicationViewController.java
@@ -16,6 +16,7 @@ import fi.otavanopisto.pyramus.dao.application.ApplicationDAO;
 import fi.otavanopisto.pyramus.dao.system.SettingDAO;
 import fi.otavanopisto.pyramus.dao.system.SettingKeyDAO;
 import fi.otavanopisto.pyramus.domainmodel.application.Application;
+import fi.otavanopisto.pyramus.domainmodel.base.Email;
 import fi.otavanopisto.pyramus.domainmodel.system.Setting;
 import fi.otavanopisto.pyramus.domainmodel.system.SettingKey;
 import fi.otavanopisto.pyramus.framework.PyramusViewController;
@@ -75,16 +76,29 @@ public class EditApplicationViewController extends PyramusViewController {
         ApplicationDAO applicationDAO = DAOFactory.getInstance().getApplicationDAO();
         Application application = applicationDAO.findByApplicationId(applicationId);
         if (application == null || application.getArchived()) {
-          pageRequestContext.getResponse().sendError(HttpServletResponse.SC_NOT_FOUND);
-          return;
+          pageRequestContext.getRequest().setAttribute("notFound", Boolean.TRUE);
+          pageRequestContext.setIncludeJSP("/templates/applications/application-edit-gateway.jsp");
         }
-        // TODO Evaluate application state; can the data still be modified by applicant?
-        pageRequestContext.getRequest().setAttribute("applicationId", applicationId);
-        pageRequestContext.getRequest().setAttribute("referenceCode", application.getReferenceCode());
-        pageRequestContext.getRequest().setAttribute("preload", Boolean.TRUE);
-        pageRequestContext.getRequest().setAttribute("donePage", Boolean.TRUE);
-        pageRequestContext.getRequest().setAttribute("saveUrl", "/1/applications/saveapplication");
-        pageRequestContext.setIncludeJSP("/templates/applications/application-edit.jsp");
+        else if (Boolean.FALSE.equals(application.getApplicantEditable())) {
+          pageRequestContext.getRequest().setAttribute("locked", Boolean.TRUE);
+          if (application.getHandler() != null) {
+            pageRequestContext.getRequest().setAttribute("handlerName", application.getHandler().getFullName());
+            Email email = application.getHandler().getPrimaryEmail();
+            if (email != null && email.getAddress() != null) {
+              pageRequestContext.getRequest().setAttribute("handlerEmail", email.getAddress());
+            }
+          }
+          pageRequestContext.setIncludeJSP("/templates/applications/application-edit-gateway.jsp");
+        }
+        else {
+          // TODO Evaluate application state; can the data still be modified by applicant?
+          pageRequestContext.getRequest().setAttribute("applicationId", applicationId);
+          pageRequestContext.getRequest().setAttribute("referenceCode", application.getReferenceCode());
+          pageRequestContext.getRequest().setAttribute("preload", Boolean.TRUE);
+          pageRequestContext.getRequest().setAttribute("donePage", Boolean.TRUE);
+          pageRequestContext.getRequest().setAttribute("saveUrl", "/1/applications/saveapplication");
+          pageRequestContext.setIncludeJSP("/templates/applications/application-edit.jsp");
+        }
       }
     }
     catch (IOException e) {
